+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * In Countly class, the old "init(directory,config)" method is deprecated, use "init(config)" instead via "instance()" call.
 * Deprecated "Countly::event" call, deprecated builder pattern. Use "Countly::events" instead.
 * Deprecated "Usage::event" call, deprecated builder pattern. Use "Countly::events" instead.
+* Deprecated "Countly::stop(boolean)" call, use "Countly::halt" instead via "instance()" call.
 
 * The following methods are deprecated from the "Event" interface:
   * "record"